refactor: centralize security utilities
